DFK-262 Fix etter CR: - Flyttet måling-testregel logikk til TestregelResource
